cli: add colors and trace level
Add a custom cijoe logger format. This is to be able to have a new
custom logging level called TRACE. The TRACE level is meant for
logging the command and it's output in the same console cijoe is being
executed and it won't be prefixed with standard format string with cijoe
caller function.

This keeps cijoe silent by default. TRACE will be enabled as first
level of debug. Add color support to differentiate between levels.

- ERROR: red
- TRACE: white
- INFO: green
- DEBUG: cyan

Source: https://stackoverflow.com/a/56944256/3638629

Signed-off-by: Daniel Gomez <da.gomez@samsung.com>